More const and other small changes. Restored save file compatibility.
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@710 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/net/general/net_happy.cpp
@@ -45,7 +45,7 @@ NetHappy::NetHappy(Unit city, Happy *data, BOOL isInitial)
 {
 	m_city = city;
 	m_data = data;
-	m_isInitialPacket = isInitial;
+	m_isInitialPacket = static_cast<uint8>(isInitial);
 }
 
 void
